Move prepare functions to form utils
